Refactored smart contract directory and sc trusted file.
