+++ blockchain/smart_contracts/sc_trusted.py
